# Speed up various functions avoiding using `jax.vmap` on indexes