--- a/tests/test_api_link.py
+++ b/tests/test_api_link.py
@@ -240,9 +240,7 @@ def test_link_bias_acceleration(
 
             W_H_L = js.model.forward_kinematics(model=model, data=data)
 
-            W_a_bias_WL = js.model.link_bias_accelerations(model=model, data=data)[
-                jnp.arange(model.number_of_links())
-            ]
+            W_a_bias_WL = js.model.link_bias_accelerations(model=model, data=data)
 
             with data.switch_velocity_representation(VelRepr.Body):
 
@@ -250,11 +248,7 @@ def test_link_bias_acceleration(
                     lambda W_H_L: jaxsim.math.Adjoint.from_transform(transform=W_H_L)
                 )(W_H_L)
 
-                L_a_bias_WL = jax.vmap(
-                    lambda index: js.link.bias_acceleration(
-                        model=model, data=data, link_index=index
-                    )
-                )(jnp.arange(model.number_of_links()))
+                L_a_bias_WL = js.model.link_bias_accelerations(model=model, data=data)
 
                 W_a_bias_WL_converted = jax.vmap(
                     lambda W_X_L, L_a_bias_WL: W_X_L @ L_a_bias_WL
@@ -267,9 +261,7 @@ def test_link_bias_acceleration(
 
             W_H_L = js.model.forward_kinematics(model=model, data=data)
 
-            L_a_bias_WL = js.model.link_bias_accelerations(model=model, data=data)[
-                jnp.arange(model.number_of_links())
-            ]
+            L_a_bias_WL = js.model.link_bias_accelerations(model=model, data=data)
 
             with data.switch_velocity_representation(VelRepr.Inertial):
 
@@ -279,9 +271,7 @@ def test_link_bias_acceleration(
                     )
                 )(W_H_L)
 
-                W_a_bias_WL = js.model.link_bias_accelerations(model=model, data=data)[
-                    jnp.arange(model.number_of_links())
-                ]
+                W_a_bias_WL = js.model.link_bias_accelerations(model=model, data=data)
 
                 L_a_bias_WL_converted = jax.vmap(
                     lambda L_X_W, W_a_bias_WL: L_X_W @ W_a_bias_WL
@@ -317,14 +307,10 @@ def test_link_jacobian_derivative(
         lambda link_index: js.link.jacobian_derivative(
             model=model, data=data, link_index=link_index
         )
-    )(js.link.names_to_idxs(model=model, link_names=model.link_names()))
+    )(jnp.arange(model.number_of_links()))
 
     # Compute the product J̇ν.
-    O_a_bias_WL = jax.vmap(
-        lambda link_index: js.link.bias_acceleration(
-            model=model, data=data, link_index=link_index
-        )
-    )(js.link.names_to_idxs(model=model, link_names=model.link_names()))
+    O_a_bias_WL = js.model.link_bias_accelerations(model=model, data=data)
 
     # Compare the two computations.
     assert jnp.einsum("l6g,g->l6", O_J̇_WL_I, I_ν) == pytest.approx(